requirements.txt
python-fasthtml
sqlite-utils
plotly
numpy
polars
xlsxwriter
fastexcel
psycopg2
requests